+++ packages/workflow-core/src/lib/plugins/external-plugin/kyc-plugin.ts
@@ -1,8 +1,7 @@
 import { AnyRecord } from '@ballerine/common';
-import { ApiPlugin } from './api-plugin';
 import { JsonSchemaValidator } from '../../utils/context-validator/json-schema-validator';
 import { Validator } from '../../utils';
-import { BallerineApiPlugin } from './ballerine-plugin';
+import { BallerineApiPlugin } from './ballerine-api-plugin';
 
 const kycIndividualRequestSchema = {
   $schema: 'http://json-schema.org/draft-07/schema#',