+++ packages/common/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/common
 
+## 0.9.70
+
+### Patch Changes
+
+- updated packages
+
+## 0.9.69
+
+### Patch Changes
+
+- updated common and core
+
+## 0.9.68
+
+### Patch Changes
+
+- version bump
+
+## 0.9.67
+
+### Patch Changes
+
+- version bump
+
 ## 0.9.66
 
 ### Patch Changes