+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/ui
 
+## 0.5.67
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.68
+
 ## 0.5.66
 
 ### Patch Changes